MP/smart-crop/ai-model/requirements.txt
flask>=3.0.2
flask-cors>=4.0.0
numpy>=1.26.4
pandas>=2.2.1
scikit-learn>=1.4.1.post1
tensorflow>=2.16.1
pillow>=10.2.0
joblib>=1.3.2
waitress>=3.0.0
